Edit request page displays some correct elements.
Not well formatted, but general request data, caller data and patient
data is displayed. The uneditable information is displayed as text and
the editable fields are input fields that can be edited.

Dropdowns are not implemented.

Questions are not yet part of the request.

This commit is just for viewing the edit request page, it doesn't add
any functionality.