Added controller for edit request (GET).
Small change made to the isLocked() method, as it seems to return true
in even if the current user is the only one attempting to edit the
request.

This currently only returns an empty request if the request is locked.
The work to be done here is doing something more sensible if the request
is locked and maybe (not sure if here or in edit request POST) actually
lock the request.